add separate lazy load image test page wip #1010
