+++ examples/app/pages/Docs.jsx
@@ -24,6 +24,7 @@ import NotFound from "./core/NotFound";
 import I18n from "./core/I18n";
 import FetchData from "./core/FetchData";
 import LazyLoadImage from "./core/LazyLoadImage";
+import LazyLoadImageDemo from "./core/LazyLoadImageDemo";
 
 // CMS
 import IntroCms from "./cms/Intro";
@@ -55,7 +56,8 @@ const PACKAGES = [
       {title: "Not Found", component: NotFound},
       {title: "I18n", component: I18n},
       {title: "FetchData", component: FetchData},
-      {title: "Lazy Load Images", component: LazyLoadImage}
+      {title: "Lazy Load Images", component: LazyLoadImage},
+      {title: "Lazy Load Images DEMO", component: LazyLoadImageDemo}
     ]
   },
   {
